pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memoryforge"
version = "0.9.0"
description = "Local-first memory layer for AI coding assistants"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "MemoryForge Team" }
]
keywords = ["ai", "memory", "mcp", "coding-assistant", "vector-database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "qdrant-client>=1.7.0",
    "mcp>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
local = [
    "sentence-transformers>=2.2.0",
]
all = [
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
]
sync = [
    "cryptography>=41.0.0",
    "gitpython>=3.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "openai>=1.0.0",
]

[project.scripts]
memoryforge = "memoryforge.cli:main"

[project.urls]
Homepage = "https://memoryforge.dev"
Repository = "https://github.com/memoryforge/core"
Documentation = "https://docs.memoryforge.dev"

[tool.hatch.build.targets.wheel]
packages = ["memoryforge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
